add baby dino to action list

+++ game-api/src/main/java/ca/ulaval/glo4002/game/application/turn/TurnUseCase.java
@@ -4,6 +4,7 @@ import ca.ulaval.glo4002.game.controllers.turn.dtos.TurnAssembler;
 import ca.ulaval.glo4002.game.domain.actions.Action;
 import ca.ulaval.glo4002.game.domain.actions.ActionRepository;
 import ca.ulaval.glo4002.game.domain.dinosaur.Dinosaur;
+import ca.ulaval.glo4002.game.domain.dinosaur.DinosaurAdult;
 import ca.ulaval.glo4002.game.domain.dinosaur.DinosaurBaby;
 import ca.ulaval.glo4002.game.domain.dinosaur.DinosaurRepository;
 import ca.ulaval.glo4002.game.domain.dinosaur.enums.DietType;
@@ -82,21 +83,23 @@ public class TurnUseCase {
     private void feedDinosaursByDietType(List<Dinosaur> sortedDinosaursByStrengthThenName) {
         boolean dead = false;
         for (Dinosaur dinosaur : sortedDinosaursByStrengthThenName) {
-            if (dinosaur.getDiet().equals(DietType.HERBIVORE)) {
-                if (!resourceRepository.eatSalad(dinosaur.feedFood())) {
+            if (dinosaur instanceof DinosaurAdult) {
+                if (dinosaur.getDiet().equals(DietType.HERBIVORE)) {
+                    if (!resourceRepository.eatSalad(dinosaur.feedFood())) {
+                        dead = true;
+                    }
+                } else {
+                    if (!resourceRepository.eatBurger(dinosaur.feedFood())) {
+                        dead = true;
+                    }
+                }
+
+                if (!resourceRepository.drinkWater(dinosaur.feedWater())) {
                     dead = true;
                 }
-            } else {
-                if (!resourceRepository.eatBurger(dinosaur.feedFood())) {
-                    dead = true;
-                }
-            }
 
-            if (!resourceRepository.drinkWater(dinosaur.feedWater())) {
-                dead = true;
+                if (dead) dinosaurRepository.remove(dinosaur);
             }
-
-            if (dead) dinosaurRepository.remove(dinosaur);
         }
 
         //Delete baby dinosaur if both parents are dead
